python/requirements.txt
# ------------------------------------------------------------
# PyTorch (torch, torchvision, torchaudio) は Apple Silicon など
# 特殊環境の場合、公式インストールガイドを参照し別途インストール
# （pip install --index-url ... や conda インストールなど）
# ------------------------------------------------------------

# Whisper (PyPI の安定版)
openai-whisper>=2023.2.2

# あるいは最新のGitHub版を使いたい場合は、下記をアンコメント
# （ただし上記の openai-whisper>=2023.2.2 と同時指定しないこと）
# git+https://github.com/openai/whisper.git

# 音源分離 Demucs
demucs>=4.0.0

# オーディオ信号処理ライブラリ
librosa>=0.10.0

# MIDI操作ライブラリ
mido>=1.2.10

# F0推定ライブラリ (必要に応じて)
crepe>=0.0.15

# ffmpeg-python (PythonからFFmpegコマンドを呼びやすくするラッパ)
# ※ 実際にはシステムにインストールされた "ffmpeg" が必要。
ffmpeg-python>=0.2.0